move custom error logic to separate file
